+++ src/main/java/com/phantomwing/eastersdelight/datagen/ModBlockStateProvider.java
@@ -2,7 +2,9 @@ package com.phantomwing.eastersdelight.datagen;
 
 import com.phantomwing.eastersdelight.EastersDelight;
 import com.phantomwing.eastersdelight.block.ModBlocks;
+import com.phantomwing.eastersdelight.block.custom.DyedEggBlock;
 import com.phantomwing.eastersdelight.block.custom.EggPainterBlock;
+import com.phantomwing.eastersdelight.component.EggPattern;
 import net.minecraft.core.registries.BuiltInRegistries;
 import net.minecraft.data.PackOutput;
 import net.minecraft.resources.ResourceLocation;
@@ -11,7 +13,6 @@ import net.minecraftforge.client.model.generators.BlockStateProvider;
 import net.minecraftforge.client.model.generators.ConfiguredModel;
 import net.minecraftforge.client.model.generators.ModelFile;
 import net.minecraftforge.common.data.ExistingFileHelper;
-import vectorwing.farmersdelight.FarmersDelight;
 
 public class ModBlockStateProvider extends BlockStateProvider {
     private static final int DEFAULT_ANGLE_OFFSET = 180;
@@ -23,6 +24,7 @@ public class ModBlockStateProvider extends BlockStateProvider {
     @Override
     protected void registerStatesAndModels() {
         eggPainterBlock(ModBlocks.EGG_PAINTER.get());
+        dyedEggBlock(ModBlocks.DYED_EGG.get());
     }
 
     private String blockName(Block block) {
@@ -37,10 +39,6 @@ public class ModBlockStateProvider extends BlockStateProvider {
         return new ModelFile.ExistingModelFile(resourceBlock(path), models().existingFileHelper);
     }
 
-    public ResourceLocation farmersDelightResourceBlock(String path) {
-        return new ResourceLocation(FarmersDelight.MODID, "block/" + path);
-    }
-
     public void eggPainterBlock(Block block) {
         getVariantBuilder(block)
                 .forAllStates(state -> ConfiguredModel.builder()
@@ -50,8 +48,26 @@ public class ModBlockStateProvider extends BlockStateProvider {
                 );
     }
 
-    public void modelBlock(Block block, String modelPath) {
-        simpleBlock(block,
-                new ModelFile.UncheckedModelFile(new ResourceLocation(EastersDelight.MOD_ID, modelPath)));
+    private void dyedEggBlock(Block block) {
+        ResourceLocation parentModel = resourceBlock(blockName(block) + "_patterned");
+
+        // Generate models for each pattern variant.
+        for (EggPattern pattern : EggPattern.values()) {
+            String modelName = blockName(block) + "_" + pattern.getName();
+            ResourceLocation modelLocation = resourceBlock(modelName);
+            this.models().withExistingParent(modelName, parentModel).texture("pattern", modelLocation).renderType("cutout");
+        }
+
+        getVariantBuilder(block)
+                .forAllStates(state -> {
+                            boolean patterned = state.getValue(DyedEggBlock.PATTERNED);
+                            EggPattern pattern = state.getValue(DyedEggBlock.EGG_PATTERN);
+
+                            return ConfiguredModel.builder()
+                                    .modelFile(existingModel(blockName(block) + (patterned ? ("_" + pattern.getName()) : "")))
+                                    .rotationY(((int) state.getValue(EggPainterBlock.FACING).toYRot() + DEFAULT_ANGLE_OFFSET) % 360)
+                                    .build();
+                        }
+                );
     }
 }