+++ src/main/java/com/phantomwing/eastersdelight/datagen/ModRecipeProvider.java
@@ -1,19 +1,14 @@
 package com.phantomwing.eastersdelight.datagen;
 
-import com.phantomwing.eastersdelight.EastersDelight;
 import com.phantomwing.eastersdelight.item.ModItems;
 import com.phantomwing.eastersdelight.tags.ForgeTags;
 import com.phantomwing.eastersdelight.tags.ModTags;
 import net.minecraft.data.PackOutput;
 import net.minecraft.data.recipes.*;
-import net.minecraft.resources.ResourceLocation;
 import net.minecraft.tags.ItemTags;
-import net.minecraft.world.item.Item;
 import net.minecraft.world.item.Items;
 import net.minecraft.world.item.crafting.*;
-import net.minecraft.world.level.ItemLike;
 import net.minecraftforge.common.crafting.conditions.IConditionBuilder;
-import net.minecraftforge.registries.RegistryObject;
 import org.jetbrains.annotations.NotNull;
 import vectorwing.farmersdelight.client.recipebook.CookingPotRecipeBookTab;
 import vectorwing.farmersdelight.data.builder.CookingPotRecipeBuilder;
@@ -69,6 +64,10 @@ public class ModRecipeProvider extends RecipeProvider implements IConditionBuild
         CuttingBoardRecipeBuilder.cuttingRecipe(Ingredient.of(ModItems.BOILED_EGG.get()), Ingredient.of(ForgeTags.TOOLS_KNIVES), ModItems.EGG_SLICE.get(), 2)
                 .addResult(Items.BONE_MEAL) // Eggshells can be used in the form of Bone Meal
                 .build(output, ModItems.EGG_SLICE.getId());
+
+        CuttingBoardRecipeBuilder.cuttingRecipe(Ingredient.of(ModItems.DYED_EGG.get()), Ingredient.of(ForgeTags.TOOLS_KNIVES), ModItems.EGG_SLICE.get(), 2)
+                .addResult(Items.BONE_MEAL) // Eggshells can be used in the form of Bone Meal
+                .build(output, ModItems.EGG_SLICE.getId() + "_from_" + ModItems.DYED_EGG.getId().getPath());
     }
 
     private void buildCookingRecipes(Consumer<FinishedRecipe> output) {
@@ -84,7 +83,7 @@ public class ModRecipeProvider extends RecipeProvider implements IConditionBuild
 
         // Farmer's Delight overrides, to include Boiled Eggs
         // Baked Cod Stew
-        CookingPotRecipeBuilder.cookingPotRecipe(vectorwing.farmersdelight.common.registry.ModItems.BAKED_COD_STEW.get(), 1, NORMAL_COOKING, MEDIUM_EXP)
+        CookingPotRecipeBuilder.cookingPotRecipe(vectorwing.farmersdelight.common.registry.ModItems.BAKED_COD_STEW.get(), 1, NORMAL_COOKING, MEDIUM_EXP, Items.BOWL)
                 .addIngredient(vectorwing.farmersdelight.common.tag.ForgeTags.RAW_FISHES_COD)
                 .addIngredient(ForgeTags.VEGETABLES_POTATO)
                 .addIngredient(ModTags.Items.BAKED_COD_STEW_INGREDIENTS)
@@ -92,96 +91,5 @@ public class ModRecipeProvider extends RecipeProvider implements IConditionBuild
                 .unlockedByAnyIngredient(Items.COD, Items.POTATO, vectorwing.farmersdelight.common.registry.ModItems.TOMATO.get(), Items.EGG)
                 .setRecipeBookTab(CookingPotRecipeBookTab.MEALS)
                 .build(output);
-    }
-
-    protected static void oneToOne(Consumer<FinishedRecipe> recipeOutput, RecipeCategory category, ItemLike item, ItemLike result, int count) {
-        ShapelessRecipeBuilder.shapeless(category, result, count)
-                .requires(item)
-                .unlockedBy(getHasName(item), has(item))
-                .save(recipeOutput, getRecipeName(item, result));
-    }
-
-    protected static void horizontalRecipe(Consumer<FinishedRecipe> recipeOutput, RecipeCategory category, ItemLike item, ItemLike result, int count) {
-        ShapedRecipeBuilder.shaped(category, result, count)
-                .pattern("###")
-                .define('#', item)
-                .unlockedBy(getHasName(item), has(item))
-                .save(recipeOutput, getRecipeName(item, result));
-    }
-
-    protected static void twoBytwo(Consumer<FinishedRecipe> recipeOutput, RecipeCategory category, ItemLike item, ItemLike result, int count) {
-        ShapedRecipeBuilder.shaped(category, result, count)
-                .pattern("##")
-                .pattern("##")
-                .define('#', item)
-                .unlockedBy(getHasName(item), has(item))
-                .save(recipeOutput, getRecipeName(item, result));
-    }
-
-    protected static void storageItemRecipes(Consumer<FinishedRecipe> recipeOutput, RecipeCategory category, ItemLike item, ItemLike storageItem) {
-        // From item to storageItem
-        ShapedRecipeBuilder.shaped(category, storageItem)
-                .pattern("###")
-                .pattern("###")
-                .pattern("###")
-                .define('#', item)
-                .unlockedBy(getHasName(item), has(item))
-                .save(recipeOutput, getRecipeName(item, storageItem));
-
-        // From storageItem to item
-        ShapelessRecipeBuilder.shapeless(category, item, 9)
-                .requires(storageItem)
-                .unlockedBy(getHasName(storageItem), has(storageItem))
-                .save(recipeOutput, getRecipeName(storageItem, item));
-    }
-
-    protected static void foodCookingRecipes(@NotNull Consumer<FinishedRecipe> recipeOutput, @NotNull ItemLike material, @NotNull ItemLike result, float experience) {
-        foodSmelting(recipeOutput, material, result, experience);
-        foodSmoking(recipeOutput, material, result, experience); // Smoking is twice as fast
-        foodCampfireCooking(recipeOutput, material, result, experience); // Campfire cooking takes three times longer
-    }
-
-    protected static void foodSmelting(@NotNull Consumer<FinishedRecipe> recipeOutput, @NotNull ItemLike material, @NotNull ItemLike result, float experience) {
-        SimpleCookingRecipeBuilder
-                .generic(Ingredient.of(material), RecipeCategory.FOOD, result, experience, 200, RecipeSerializer.SMELTING_RECIPE)
-                .unlockedBy(getHasName(material), has(material))
-                .save(recipeOutput);
-    }
-
-    protected static void foodSmoking(@NotNull Consumer<FinishedRecipe> recipeOutput, @NotNull ItemLike material, @NotNull ItemLike result, float experience) {
-        SimpleCookingRecipeBuilder
-                .generic(Ingredient.of(material), RecipeCategory.FOOD, result, experience, 100, RecipeSerializer.SMOKING_RECIPE)
-                .unlockedBy(getHasName(material), has(material))
-                .save(recipeOutput, EastersDelight.MOD_ID + ":" + getItemName(result) + "_from_smoking");
-    }
-
-    protected static void foodCampfireCooking(@NotNull Consumer<FinishedRecipe> recipeOutput, @NotNull ItemLike material, @NotNull ItemLike result, float experience) {
-        SimpleCookingRecipeBuilder
-                .generic(Ingredient.of(material), RecipeCategory.FOOD, result, experience, 600, RecipeSerializer.CAMPFIRE_COOKING_RECIPE)
-                .unlockedBy(getHasName(material), has(material))
-                .save(recipeOutput, EastersDelight.MOD_ID + ":" + getItemName(result) + "_from_campfire_cooking");
-    }
-
-    protected static void pieRecipes(Consumer<FinishedRecipe> recipeOutput, @NotNull RegistryObject<Item> pieBlock, @NotNull RegistryObject<Item> sliceItem, Ingredient topping) {
-        ShapedRecipeBuilder.shaped(RecipeCategory.FOOD, pieBlock.get(), 1)
-                .pattern("TTT")
-                .pattern("MMM")
-                .pattern("SCS")
-                .define('T', topping)
-                .define('M', ForgeTags.MILK)
-                .define('S', Items.SUGAR)
-                .define('C', vectorwing.farmersdelight.common.registry.ModItems.PIE_CRUST.get())
-                .unlockedBy(getHasName(vectorwing.farmersdelight.common.registry.ModItems.PIE_CRUST.get()), has(vectorwing.farmersdelight.common.registry.ModItems.PIE_CRUST.get()))
-                .save(recipeOutput);
-        ShapedRecipeBuilder.shaped(RecipeCategory.FOOD, pieBlock.get(), 1)
-                .pattern("##")
-                .pattern("##")
-                .define('#', sliceItem.get())
-                .unlockedBy(getHasName(sliceItem.get()), has(sliceItem.get()))
-                .save(recipeOutput, new ResourceLocation(EastersDelight.MOD_ID, getItemName(pieBlock.get()) + "_from_slices"));
-    }
-
-    protected static String getRecipeName(ItemLike item, ItemLike result) {
-        return EastersDelight.MOD_ID + ":" + getConversionRecipeName(result, item);
     }
 }